itests: Breakup ccupgrade_test.go
